[#IOPAE-586] add InfoIconWithTooltip reusable component
