+++ src/components/subscriptions/ApiKey.css
@@ -4,12 +4,3 @@
   margin-top: 0px;
   font-size: 16px;
 }
-
-.api-key--info {
-  margin-left: 4px;
-  vertical-align: text-bottom;
-}
-
-.pointer {
-  cursor: pointer;
-}
