+++ src/components/subscriptions/InfoIconWithTooltip.css
@@ -1,0 +1,8 @@
+.info-icon--info {
+    margin-left: 4px;
+    vertical-align: text-bottom;
+  }
+  
+  .pointer {
+    cursor: pointer;
+  }
